Autoload classmap from src/
No longer using classes.php and dbcon.php to store classes. All classes now enforce strict types. Will opt for PSR-4 autoloading in the future - classmap autloading should do the job for now.
